renames importHTML to ts
